[Refactor]: PostPage 관련 컴포넌트 위치 조정
