+++ client/src/feature/post/ui/PostCard/PostContent/index.jsx
